requirements.txt
# Core dependencies
numpy>=1.20.0
pillow>=8.0.0
opencv-python>=4.5.0
matplotlib>=3.3.0
tqdm>=4.60.0

# Data processing
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning / Deep Learning
torch>=1.9.0
transformers>=4.20.0
peft>=0.3.0

# OpenAI API
openai>=1.0.0
python-dotenv>=0.19.0

# Qwen VL utilities
qwen-vl-utils>=0.1.0

# YAML processing
pyyaml>=5.4.0

# Testing (optional, for development)
pytest>=7.0.0
pytest-cov>=4.0.0

# Image processing
rasterio>=1.2.0  # Optional, for TIF file handling

# Additional utilities
requests>=2.25.0

# Note: The following modules may need to be provided separately:
# - modality_utils (image_converters, path_resolver, messaging)
# - qwen3_svf_qa (fallback if modality_utils is not available)
# - scene_statistics (for freeform QA generation)
# - validation.question_validators (for question validation)

